Some tweaks to the starter webapp
