+++ src/main/resources/static/styles/general.css
@@ -1,5 +1,3 @@
-
-
 * {
     box-sizing: border-box;
 }
