+++ src/main/resources/static/styles/navbar.css
@@ -1,6 +1,6 @@
-
 nav {
     width: 100%;
+    flex-shrink: 0;
 
     display: flex;
     flex-direction: row;
@@ -9,7 +9,6 @@ nav {
     background-color: #343a40;
     padding: 15px 0;
 }
-
 
 .navbar-title {
     margin: 0 5px 0 20px;
