+++ src/main/resources/templates/about.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
 
+    <!--@thymesVar id="model" type="org.softwire.training.bookish.models.page.AboutPageModel"-->
 
     <head>
         <title>Bookish - About Us</title>
@@ -10,11 +11,10 @@
         <link rel="stylesheet" href="/styles/about.css"/>
     </head>
 
-
     <body>
 
         <div th:replace="fragments/navbar :: navbar">
-            <!-- This element is replaced the the HTML in fragments/navbar.html -->
+            <!-- This element is replaced by the HTML in fragments/navbar.html -->
         </div>
 
         <div class="container">
