+++ src/main/resources/templates/index.html
@@ -10,11 +10,10 @@
         <link rel="stylesheet" href="/styles/index.css"/>
     </head>
 
-
     <body>
 
         <div th:replace="fragments/navbar :: navbar">
-            <!-- This element is replaced the the HTML in fragments/navbar.html -->
+            <!-- This element is replaced by the HTML in fragments/navbar.html -->
         </div>
 
         <div class="container">
